pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Van1s1mys <malkovsin@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}, {include = "services"}, {include = "utils"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
langchain = "^0.2.0"
langchain-community = "^0.2.0"
chromadb = { version = "^0.6.3", extras = ["http"] }
aiohttp = "^3.9.0"
python-multipart = "0.0.19"
pypdf = "^4.2.0"
beautifulsoup4 = "^4.13.3"
psutil = "^7.0.0"
langchain-chroma = "^0.2.2"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"